Fix uint8_t counter overflow in node_info DNS address tracking
resolved_address_count and current_address were uint8_t, limiting
DNS resolution to 255 addresses. If a hostile DNS server returns
more than 255 records, the counter wraps to 0, causing g_realloc
to shrink the buffer and subsequent writes to corrupt the heap.
Changed to uint16_t which supports up to 65535 addresses.

+++ include/node_info.h
@@ -32,8 +32,8 @@ struct node_info
 	struct
 	{
 		char **resolved_addresses;
-		uint8_t resolved_address_count;
-		uint8_t current_address;
+		uint16_t resolved_address_count;
+		uint16_t current_address;
 
 		struct dns_query *a;
 		struct dns_query *aaaa;
